+++ services/web/server/src/simcore_service_webserver/projects/projects_api.py
@@ -182,13 +182,13 @@ async def start_project_interactive_services(
 async def delete_project(app: web.Application, project_uuid: str, user_id: int) -> None:
     await delete_project_from_db(app, project_uuid, user_id)
 
-    async def remove_services_and_data():
+    async def _remove_services_and_data():
         await remove_project_interactive_services(
             user_id, project_uuid, app, notify_users=False
         )
         await delete_project_data(app, project_uuid, user_id)
 
-    fire_and_forget_task(remove_services_and_data())
+    fire_and_forget_task(_remove_services_and_data())
 
 
 @observe(event="SIGNAL_USER_DISCONNECTED")
