+++ services/web/server/tests/unit/with_dbs/09/test_projects_crud.py
@@ -3,233 +3,28 @@
 # pylint: disable=unused-argument
 # pylint: disable=unused-variable
 
-import asyncio
+
 import uuid as uuidlib
 from copy import deepcopy
 from math import ceil
 from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union
-from unittest.mock import call
 
 import pytest
-from _helpers import ExpectedResponse, standard_role_response
+from _helpers import standard_role_response
 from aiohttp import web
 from aiohttp.test_utils import TestClient
 from aioresponses import aioresponses
-from models_library.projects_access import Owner
-from models_library.projects_state import (
-    ProjectLocked,
-    ProjectRunningState,
-    ProjectState,
-    ProjectStatus,
-    RunningState,
-)
+from models_library.projects_state import ProjectState
 from pytest_simcore.helpers.utils_assert import assert_status
-from pytest_simcore.helpers.utils_projects import NewProject, delete_all_projects
-from servicelib import async_utils
-from servicelib.aiohttp.application import create_safe_application
-from simcore_service_webserver import catalog
-from simcore_service_webserver.db import setup_db
+from simcore_service_webserver._meta import api_version_prefix
 from simcore_service_webserver.db_models import UserRole
-from simcore_service_webserver.director.module_setup import setup_director
-from simcore_service_webserver.director_v2 import setup_director_v2
-from simcore_service_webserver.login.module_setup import setup_login
-from simcore_service_webserver.products import setup_products
-from simcore_service_webserver.projects.module_setup import setup_projects
 from simcore_service_webserver.projects.projects_handlers import (
     OVERRIDABLE_DOCUMENT_KEYS,
 )
-from simcore_service_webserver.resource_manager.module_setup import (
-    setup_resource_manager,
-)
-from simcore_service_webserver.rest import setup_rest
-from simcore_service_webserver.security import setup_security
-from simcore_service_webserver.session import setup_session
-from simcore_service_webserver.socketio.module_setup import setup_socketio
-from simcore_service_webserver.tags import setup_tags
 from simcore_service_webserver.utils import now_str, to_datetime
-from socketio.exceptions import ConnectionError as SocketConnectionError
 from yarl import URL
 
-API_VERSION = "v0"
-RESOURCE_NAME = "projects"
-API_PREFIX = "/" + API_VERSION
-
-
-DEFAULT_GARBAGE_COLLECTOR_INTERVAL_SECONDS: int = 3
-DEFAULT_GARBAGE_COLLECTOR_DELETION_TIMEOUT_SECONDS: int = 3
-
-
-@pytest.fixture
-def client(
-    loop,
-    aiohttp_client,
-    app_cfg,
-    postgres_db,
-    mocked_director_v2_api,
-    mock_orphaned_services,
-    redis_client,
-):
-
-    # config app
-    cfg = deepcopy(app_cfg)
-    port = cfg["main"]["port"]
-    cfg["projects"]["enabled"] = True
-    cfg["director"]["enabled"] = True
-    cfg["resource_manager"][
-        "garbage_collection_interval_seconds"
-    ] = DEFAULT_GARBAGE_COLLECTOR_INTERVAL_SECONDS  # increase speed of garbage collection
-    cfg["resource_manager"][
-        "resource_deletion_timeout_seconds"
-    ] = DEFAULT_GARBAGE_COLLECTOR_DELETION_TIMEOUT_SECONDS  # reduce deletion delay
-    app = create_safe_application(cfg)
-
-    # setup app
-    setup_db(app)
-    setup_session(app)
-    setup_security(app)
-    setup_rest(app)
-    setup_login(app)  # needed for login_utils fixtures
-    setup_resource_manager(app)
-    setup_socketio(app)
-    setup_director(app)
-    setup_director_v2(app)
-    setup_tags(app)
-    assert setup_projects(app)
-    setup_products(app)
-
-    # server and client
-    yield loop.run_until_complete(
-        aiohttp_client(app, server_kwargs={"port": port, "host": "localhost"})
-    )
-
-    # teardown here ...
-
-
-@pytest.fixture()
-def ensure_run_in_sequence_context_is_empty():
-    async_utils.sequential_jobs_contexts = {}
-
-
-@pytest.fixture
-def mocks_on_projects_api(mocker, logged_user) -> None:
-    """
-    All projects in this module are UNLOCKED
-
-    Emulates that it found logged_user as the SOLE user of this project
-    and returns the  ProjectState indicating his as owner
-    """
-    nameparts = logged_user["name"].split(".") + [""]
-    state = ProjectState(
-        locked=ProjectLocked(
-            value=False,
-            owner=Owner(
-                user_id=logged_user["id"],
-                first_name=nameparts[0],
-                last_name=nameparts[1],
-            ),
-            status=ProjectStatus.CLOSED,
-        ),
-        state=ProjectRunningState(value=RunningState.NOT_STARTED),
-    )
-    mocker.patch(
-        "simcore_service_webserver.projects.projects_api._get_project_lock_state",
-        return_value=state,
-    )
-
-
-@pytest.fixture
-async def user_project(client, fake_project, logged_user):
-    async with NewProject(
-        fake_project, client.app, user_id=logged_user["id"]
-    ) as project:
-        print("-----> added project", project["name"])
-        yield project
-        print("<----- removed project", project["name"])
-
-
-@pytest.fixture
-async def shared_project(client, fake_project, logged_user, all_group):
-    fake_project.update(
-        {
-            "accessRights": {
-                f"{all_group['gid']}": {"read": True, "write": False, "delete": False}
-            },
-        },
-    )
-    async with NewProject(
-        fake_project,
-        client.app,
-        user_id=logged_user["id"],
-    ) as project:
-        print("-----> added project", project["name"])
-        yield project
-        print("<----- removed project", project["name"])
-
-
-@pytest.fixture
-async def template_project(
-    client, fake_project, logged_user, all_group: Dict[str, str]
-):
-    project_data = deepcopy(fake_project)
-    project_data["name"] = "Fake template"
-    project_data["uuid"] = "d4d0eca3-d210-4db6-84f9-63670b07176b"
-    project_data["accessRights"] = {
-        str(all_group["gid"]): {"read": True, "write": False, "delete": False}
-    }
-
-    async with NewProject(
-        project_data, client.app, user_id=None, clear_all=True
-    ) as template_project:
-        print("-----> added template project", template_project["name"])
-        yield template_project
-        print("<----- removed template project", template_project["name"])
-
-
-@pytest.fixture
-def fake_services():
-    def create_fakes(number_services: int) -> List[Dict]:
-        fake_services = [{"service_uuid": f"{i}_uuid"} for i in range(number_services)]
-        return fake_services
-
-    yield create_fakes
-
-
-@pytest.fixture
-async def project_db_cleaner(client):
-    yield
-    await delete_all_projects(client.app)
-
-
-@pytest.fixture
-async def catalog_subsystem_mock(
-    monkeypatch,
-) -> Callable[[Optional[Union[List[Dict], Dict]]], None]:
-    services_in_project = []
-
-    def creator(projects: Optional[Union[List[Dict], Dict]] = None) -> None:
-        for proj in projects:
-            services_in_project.extend(
-                [
-                    {"key": s["key"], "version": s["version"]}
-                    for _, s in proj["workbench"].items()
-                ]
-            )
-
-    async def mocked_get_services_for_user(*args, **kwargs):
-        return services_in_project
-
-    monkeypatch.setattr(
-        catalog, "get_services_for_user_in_product", mocked_get_services_for_user
-    )
-
-    return creator
-
-
-@pytest.fixture(autouse=True)
-async def director_v2_automock(
-    director_v2_service_mock: aioresponses,
-) -> aioresponses:
-    yield director_v2_service_mock
+API_PREFIX = "/" + api_version_prefix
 
 
 # HELPERS -----------------------------------------------------------------------------------------
@@ -450,15 +245,6 @@ async def _replace_project(
     if not error:
         assert_replaced(current_project=data, update_data=project_update)
     return data
-
-
-async def _delete_project(
-    client, project: Dict, expected: Type[web.HTTPException]
-) -> None:
-    url = client.app.router["delete_project"].url_for(project_id=project["uuid"])
-    assert str(url) == f"{API_PREFIX}/projects/{project['uuid']}"
-    resp = await client.delete(url)
-    await assert_status(resp, expected)
 
 
 # TESTS ----------------------------------------------------------------------------------------------------
@@ -838,187 +624,3 @@ async def test_replace_project_updated_readonly_inputs(
         "Kr"
     ] = 5
     await _replace_project(client, project_update, expected)
-
-
-# DELETE -------
-@pytest.mark.parametrize(
-    "user_role,expected",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPForbidden),
-        (UserRole.USER, web.HTTPNoContent),
-        (UserRole.TESTER, web.HTTPNoContent),
-    ],
-)
-async def test_delete_project(
-    client,
-    logged_user,
-    user_project,
-    expected,
-    storage_subsystem_mock,
-    mocked_director_v2_api,
-    catalog_subsystem_mock,
-    fake_services,
-):
-    # DELETE /v0/projects/{project_id}
-
-    fakes = fake_services(5)
-    mocked_director_v2_api["director_v2_api.get_services"].return_value = fakes
-
-    await _delete_project(client, user_project, expected)
-    await asyncio.sleep(2)  # let some time fly for the background tasks to run
-
-    if expected == web.HTTPNoContent:
-        mocked_director_v2_api["director_v2_api.get_services"].assert_called_once()
-
-        expected_calls = [
-            call(
-                app=client.server.app,
-                service_uuid=service["service_uuid"],
-                save_state=True,
-            )
-            for service in fakes
-        ]
-        mocked_director_v2_api["director_v2_api.stop_service"].assert_has_calls(
-            expected_calls
-        )
-
-        # wait for the fire&forget to run
-        await asyncio.sleep(2)
-        await _assert_get_same_project(client, user_project, web.HTTPNotFound)
-
-
-@pytest.mark.parametrize(
-    "user_role, expected_ok, expected_forbidden",
-    [
-        (UserRole.ANONYMOUS, web.HTTPUnauthorized, web.HTTPUnauthorized),
-        (UserRole.GUEST, web.HTTPOk, web.HTTPForbidden),
-        (UserRole.USER, web.HTTPOk, web.HTTPForbidden),
-        (UserRole.TESTER, web.HTTPOk, web.HTTPForbidden),
-    ],
-)
-async def test_delete_multiple_opened_project_forbidden(
-    client,
-    logged_user,
-    user_project,
-    mocked_director_v2_api,
-    mocked_dynamic_service,
-    socketio_client_factory: Callable,
-    client_session_id_factory: Callable,
-    user_role,
-    expected_ok,
-    expected_forbidden,
-):
-    # service in project = await mocked_dynamic_service(logged_user["id"], empty_user_project["uuid"])
-    service = await mocked_dynamic_service(logged_user["id"], user_project["uuid"])
-    # open project in tab1
-    client_session_id1 = client_session_id_factory()
-    try:
-        sio1 = await socketio_client_factory(client_session_id1)
-    except SocketConnectionError:
-        if user_role != UserRole.ANONYMOUS:
-            pytest.fail("socket io connection should not fail")
-    url = client.app.router["open_project"].url_for(project_id=user_project["uuid"])
-    resp = await client.post(url, json=client_session_id1)
-    await assert_status(resp, expected_ok)
-    # delete project in tab2
-    client_session_id2 = client_session_id_factory()
-    try:
-        sio2 = await socketio_client_factory(client_session_id2)
-    except SocketConnectionError:
-        if user_role != UserRole.ANONYMOUS:
-            pytest.fail("socket io connection should not fail")
-    await _delete_project(client, user_project, expected_forbidden)
-
-
-# PAGINATION
-def standard_user_role() -> Tuple[str, Tuple[UserRole, ExpectedResponse]]:
-    all_roles = standard_role_response()
-
-    return (all_roles[0], [pytest.param(*all_roles[1][2], id="standard user role")])
-
-
-@pytest.mark.parametrize(
-    "limit, offset, expected_error_msg",
-    [
-        (-7, 0, "Invalid parameter value for `limit`"),
-        (0, 0, "Invalid parameter value for `limit`"),
-        (43, -2, "Invalid parameter value for `offset`"),
-    ],
-)
-@pytest.mark.parametrize(*standard_user_role())
-async def test_list_projects_with_invalid_pagination_parameters(
-    client: TestClient,
-    logged_user: Dict[str, Any],
-    primary_group: Dict[str, str],
-    expected: ExpectedResponse,
-    storage_subsystem_mock,
-    catalog_subsystem_mock: Callable[[Optional[Union[List[Dict], Dict]]], None],
-    director_v2_service_mock: aioresponses,
-    project_db_cleaner,
-    limit: int,
-    offset: int,
-    expected_error_msg: str,
-):
-    await _list_projects(
-        client,
-        web.HTTPBadRequest,
-        query_parameters={"limit": limit, "offset": offset},
-        expected_error_msg=expected_error_msg,
-        expected_error_code="InvalidParameterValue",
-    )
-
-
-@pytest.mark.parametrize("limit", [7, 20, 43])
-@pytest.mark.parametrize(*standard_user_role())
-async def test_list_projects_with_pagination(
-    client: TestClient,
-    logged_user: Dict[str, Any],
-    primary_group: Dict[str, str],
-    expected: ExpectedResponse,
-    storage_subsystem_mock,
-    catalog_subsystem_mock: Callable[[Optional[Union[List[Dict], Dict]]], None],
-    director_v2_service_mock: aioresponses,
-    project_db_cleaner,
-    limit: int,
-):
-
-    NUM_PROJECTS = 90
-    # let's create a few projects here
-    created_projects = await asyncio.gather(
-        *[
-            _new_project(client, expected.created, logged_user, primary_group)
-            for i in range(NUM_PROJECTS)
-        ]
-    )
-    if expected.created == web.HTTPCreated:
-        catalog_subsystem_mock(created_projects)
-
-        assert len(created_projects) == NUM_PROJECTS
-        NUMBER_OF_CALLS = ceil(NUM_PROJECTS / limit)
-        next_link = None
-        default_query_parameter = {"limit": limit}
-        projects = []
-        for i in range(NUMBER_OF_CALLS):
-            print(
-                "calling in with query",
-                next_link.query if next_link else default_query_parameter,
-            )
-            data, meta, links = await _list_projects(
-                client,
-                expected.ok,
-                query_parameters=next_link.query
-                if next_link
-                else default_query_parameter,
-            )
-            print("...received [", meta, "]")
-            assert len(data) == meta["count"]
-            assert meta["count"] == min(limit, NUM_PROJECTS - len(projects))
-            assert meta["limit"] == limit
-            projects.extend(data)
-            next_link = URL(links["next"]) if links["next"] is not None else None
-
-        assert len(projects) == len(created_projects)
-        assert {prj["uuid"] for prj in projects} == {
-            prj["uuid"] for prj in created_projects
-        }
